결제 API / 장바구니 수정

+++ src/main/java/com/kh/adminGoods/AdminGoodsController.java
@@ -123,7 +123,7 @@ public class AdminGoodsController {
 
 		mv.addObject("goodsList", goodsList);
 		Integer count = adminGoodsService.goodsCount();
-		int a = (int) count;
+		int a = count;
 		mv.addObject("count", a);
 
 		return mv;
